Add new testcases for nats source
